Move stat calc scripts to angular service
