+++ karma.conf.js
@@ -55,6 +55,7 @@ module.exports = function(config) {
       "website/public/js/services/taskServices.js",
       "website/public/js/services/paymentServices.js",
       "website/public/js/services/questServices.js",
+      "website/public/js/services/statServices.js",
 
       "website/public/js/filters/money.js",
       "website/public/js/filters/roundLargeNumbers.js",
