+++ website/public/manifest.json
@@ -52,6 +52,7 @@
             "js/services/challengeServices.js",
             "js/services/paymentServices.js",
             "js/services/questServices.js",
+            "js/services/statServices.js",
 
             "js/filters/money.js",
             "js/filters/roundLargeNumbers.js",
